COMMON: move bidi support to str.h and ustr.h
