+++ common/translation.cpp
@@ -33,7 +33,6 @@
 #include "common/fs.h"
 #include "common/system.h"
 #include "common/textconsole.h"
-#include "common/str-bidi.h"
 
 #ifdef USE_TRANSLATION
